+++ src/ass/__init__.py
@@ -1,3 +1,3 @@
 from .from_ass import read_karaoke
-from .to_ass import effects
+from .to_ass import register_effect, retrieve_effect
 from .consts import *
