+++ src/effect/__init__.py
@@ -1,0 +1,3 @@
+from .default_effect import *
+from .no_karaoke_effect import *
+from .no_lyrics_effect import *
