+++ src/effect/default_effect.py
@@ -1,0 +1,92 @@
+from collections.abc import Sequence, Mapping
+from datetime import timedelta
+
+import pyass
+
+from ..ass.consts import *
+from ..ass.to_ass import *
+from ..models.karaoke import *
+
+
+def to_default_event(
+    line: KLine,
+    actorToStyle: Mapping[str, Sequence[pyass.Tag]],
+    switchDuration: timedelta,
+    transitionDuration: timedelta,
+) -> pyass.Event:
+    if line.isEN:
+        line.calculate_char_offsets(EN_STYLE, transitionDuration)
+    else:
+        line.calculate_char_offsets(ROMAJI_STYLE, transitionDuration)
+
+    # Generate line style tags
+    eventParts: list[pyass.EventPart] = [
+        pyass.EventPart(
+            tags=[
+                *LYRICS_TAGS,
+                get_en_pos_tag(line) if line.isEN else get_romaji_pos_tag(line),
+            ],
+        )
+    ]
+
+    # Set color if there's a constant actor
+    if len(line.actorSwitches) == 0:
+        eventParts.append(pyass.EventPart(tags=actorToStyle[line.startActor]))
+
+    # Leading switch duration
+    if not line.isEN:
+        eventParts.append(pyass.EventPart(tags=[pyass.KaraokeTag(switchDuration)]))
+
+    # Generate character style tags
+    for char in line.chars:
+        eventParts.append(
+            pyass.EventPart(
+                tags=[
+                    *get_char_transform_tags(char, switchDuration, transitionDuration),
+                    *get_char_actor_tag(char, actorToStyle, switchDuration),
+                    *get_char_karaoke_tag(char),
+                ],
+                text=char.char,
+            )
+        )
+
+    # Trailing switch duration
+    if not line.isEN:
+        eventParts.append(pyass.EventPart(tags=[pyass.KaraokeTag(switchDuration)]))
+
+    return pyass.Event(
+        format=get_line_format(line),
+        style=EN_STYLE.name if line.isEN else ROMAJI_STYLE.name,
+        start=line.start - switchDuration,
+        end=line.end + switchDuration,
+        parts=eventParts,
+    )
+
+
+class DefaultLiveKaraokeEffect(KaraokeEffect):
+    def to_romaji_k_events(
+        self,
+        songLines: Sequence[KLine],
+        actorToStyle: Mapping[str, Sequence[pyass.Tag]],
+    ) -> Sequence[pyass.Event]:
+        return [
+            to_default_event(
+                line, actorToStyle, DEFAULT_SWITCH_DURATION, DEFAULT_TRANSITION_DURATION
+            )
+            for line in songLines
+        ]
+
+    def to_en_k_events(
+        self,
+        songLines: Sequence[KLine],
+        actorToStyle: Mapping[str, Sequence[pyass.Tag]],
+    ) -> Sequence[pyass.Event]:
+        return [
+            to_default_event(
+                line, actorToStyle, DEFAULT_SWITCH_DURATION, DEFAULT_TRANSITION_DURATION
+            )
+            for line in songLines
+        ]
+
+
+register_effect("default_live_karaoke_effect", DefaultLiveKaraokeEffect())
